requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.10.0
python-multipart==0.0.6
chromadb==0.4.22
pypdf2==3.0.1
python-docx==1.1.0
pydantic==2.5.0
httpx==0.26.0
pandas==2.1.4
openpyxl==3.1.2
gunicorn==21.2.0
werkzeug==3.0.1
